ADD: server.js 추가
